move checkout to its own file
